requirements.txt
parsedatetime
google-auth
google-auth-oauthlib
google-api-python-client
pyannote.audio
python-dotenv
google-generativeai
Flask
sounddevice
numpy
openai-whisper
soundfile
dateparser
